+++ lib/dorcet/yoghet.dart
@@ -24,6 +24,6 @@ class Yoghet extends Dorcet {
     }
 
     final pposition = Vector2(x + offset, y);
-    game.add(Yoghectile(speed: speed, damage: damage, position: pposition));
+    game.add(Yoghectile(speed: speed, damage: damage.toDouble(), position: pposition));
   }
 }
